refactor(singbox,storage): разнос operator.go и settings.go по темам — чистые перемещения
Этап A реструктуризации backend:

internal/singbox/operator.go (3431 строка, 116 деклараций) разнесён по
байтовым AST-диапазонам (декларации перенесены дословно, включая
doc-комментарии):
- operator_process.go — жизненный цикл процесса: start/stop/reload,
  разбор stderr/stdout, OOM-детект по dmesg, crash-статистика,
  автоперезапуск, ручной стоп, ожидание Clash
- operator_baseconfig.go — патч-конвейер 00-base.json: ensureBaseConfig,
  patchBase*/filterOut*, миграция легаси-конфига, fresh-base
- operator_install.go — версия/фичи бинаря (probe + sidecar-кэш),
  install/update, GetStatus
- operator_tunnels.go — CRUD пользовательских туннелей: парсинг
  share-ссылок, аллокация тегов/портов, ApplyConfig, селекторы
- operator_ndmsproxy.go — orphan-cleanup NDMS Proxy-интерфейсов
- operator.go (ядро, 436 строк) — тип Operator, Deps, конструктор,
  сеттеры проводки, аксессоры

internal/storage/settings.go: 32 миграции (migrateToV2..V30,
migrateNATModes, migrateManagedServers, migratePortFile) вынесены в
settings_migrations.go; в settings.go остались Load/Save/типизированные
аксессоры.

Эквивалентность верифицирована мультимножествами нормализованных строк
кода И комментариев (старый файл против суммы новых): расхождений нет.
Поведение не меняется.
